# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Esclavizador",
    "description": "Corporate time tracking utility with multi-tenant support",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create project",
        "description": "Create new project (boss only)",
        "operationId": "create_project_api_v1_projects_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List projects",
        "description": "List all projects in user's organization with optional filtering",
        "operationId": "list_projects_api_v1_projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "title": "Is Active"
            },
            "description": "Filter by active status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum items per page",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create task",
        "description": "Create new task (boss only)",
        "operationId": "create_task_api_v1_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "List tasks",
        "description": "List all tasks in user's organization with optional filtering",
        "operationId": "list_tasks_api_v1_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by project ID",
              "title": "Project Id"
            },
            "description": "Filter by project ID"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "title": "Is Active"
            },
            "description": "Filter by active status"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum items per page",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/time-entries": {
      "post": {
        "tags": [
          "Time Entries"
        ],
        "summary": "Create manual entry",
        "description": "Create a manual time entry (already completed)",
        "operationId": "create_manual_entry_api_v1_time_entries_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Time Entries"
        ],
        "summary": "List time entries",
        "description": "List time entries with optional filtering (workers see own, bosses see all)",
        "operationId": "list_time_entries_api_v1_time_entries_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by project ID",
              "title": "Project Id"
            },
            "description": "Filter by project ID"
          },
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by task ID",
              "title": "Task Id"
            },
            "description": "Filter by task ID"
          },
          {
            "name": "is_billable",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by billable status",
              "title": "Is Billable"
            },
            "description": "Filter by billable status"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by user ID (bosses only)",
              "title": "User Id"
            },
            "description": "Filter by user ID (bosses only)"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by start date (entries >= this date)",
              "title": "Start Date"
            },
            "description": "Filter by start date (entries >= this date)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by end date (entries <= this date)",
              "title": "End Date"
            },
            "description": "Filter by end date (entries <= this date)"
          },
          {
            "name": "is_running",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by running status",
              "title": "Is Running"
            },
            "description": "Filter by running status"
          },
          {
            "name": "tag_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag IDs (OR logic)",
              "title": "Tag Ids"
            },
            "description": "Filter by tag IDs (OR logic)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum items per page",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeEntryList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tags": {
      "post": {
        "tags": [
          "Tags"
        ],
        "summary": "Create tag",
        "description": "Create new tag (boss only)",
        "operationId": "create_tag_api_v1_tags_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tags"
        ],
        "summary": "List tags",
        "description": "List all tags in user's organization",
        "operationId": "list_tags_api_v1_tags_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum items per page",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "description": "Create new user in organization (boss only)",
        "operationId": "create_user_api_v1_users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "description": "List all users in organization (boss only)",
        "operationId": "list_users_api_v1_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "title": "Is Active"
            },
            "description": "Filter by active status"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by role (boss/worker)",
              "title": "Role"
            },
            "description": "Filter by role (boss/worker)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum items per page",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum items per page"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Name",
            "description": "Project name (2-255 characters)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Project description (optional)"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "description": "Schema for creating a new project.",
        "example": {
          "description": "Complete redesign of company website",
          "name": "Website Redesign"
        }
      },
      "ProjectList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of projects"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of projects matching filters"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Maximum items per page"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Number of items skipped"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ProjectList",
        "description": "Schema for paginated project list response.",
        "example": {
          "items": [
            {
              "created_at": "2025-01-15T12:00:00Z",
              "description": "Complete redesign of company website",
              "id": "123e4567-e89b-12d3-a456-426614174000",
              "is_active": true,
              "name": "Website Redesign",
              "organization_id": "123e4567-e89b-12d3-a456-426614174001",
              "task_count": 5
            }
          ],
          "limit": 50,
          "offset": 0,
          "total": 1
        }
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Project unique identifier"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Project name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Project description"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id",
            "description": "Organization ID"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Active status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "task_count": {
            "type": "integer",
            "title": "Task Count",
            "description": "Number of tasks in this project",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "organization_id",
          "is_active",
          "created_at"
        ],
        "title": "ProjectResponse",
        "description": "Schema for project response (used for all endpoints).",
        "example": {
          "created_at": "2025-01-15T12:00:00Z",
          "description": "Complete redesign of company website",
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "is_active": true,
          "name": "Website Redesign",
          "organization_id": "123e4567-e89b-12d3-a456-426614174001",
          "task_count": 5
        }
      },
...
      "TaskCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Name",
            "description": "Task name (2-255 characters)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Task description (optional)"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id",
            "description": "Project ID this task belongs to"
          }
        },
        "type": "object",
        "required": [
          "name",
          "project_id"
        ],
        "title": "TaskCreate",
        "description": "Schema for creating a new task.",
        "example": {
          "description": "Create initial design mockups in Figma",
          "name": "Design mockups",
          "project_id": "123e4567-e89b-12d3-a456-426614174000"
        }
      },
      "TaskList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "List of tasks"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of tasks matching filters"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Maximum items per page"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Number of items skipped"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "TaskList",
        "description": "Schema for paginated task list response.",
        "example": {
          "items": [
            {
              "created_at": "2025-01-15T12:00:00Z",
              "description": "Create initial design mockups in Figma",
              "id": "123e4567-e89b-12d3-a456-426614174002",
              "is_active": true,
              "name": "Design mockups",
              "project_id": "123e4567-e89b-12d3-a456-426614174000",
              "project_name": "Website Redesign"
            }
          ],
          "limit": 50,
          "offset": 0,
          "total": 1
        }
      },
      "TaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Task unique identifier"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Task name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Task description"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id",
            "description": "Project ID"
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "description": "Project name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Active status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "project_id",
          "project_name",
          "is_active",
          "created_at"
        ],
        "title": "TaskResponse",
        "description": "Schema for task response (used for all endpoints).",
        "example": {
          "created_at": "2025-01-15T12:00:00Z",
          "description": "Create initial design mockups in Figma",
          "id": "123e4567-e89b-12d3-a456-426614174002",
          "is_active": true,
          "name": "Design mockups",
          "project_id": "123e4567-e89b-12d3-a456-426614174000",
          "project_name": "Website Redesign"
        }
      },
...
      "TimeEntryCreate": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id",
            "description": "Project to track time for"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id",
            "description": "Optional task within project"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time",
            "description": "Entry start time (UTC)"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time",
            "description": "Entry end time (UTC)"
          },
          "is_billable": {
            "type": "boolean",
            "title": "Is Billable",
            "description": "Whether time is billable",
            "default": true
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Entry description"
          },
          "tag_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Tag Ids",
            "description": "Tag IDs to assign"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "start_time",
          "end_time"
        ],
        "title": "TimeEntryCreate",
        "description": "Schema for creating a manual time entry."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}